prodegeit/requirements.txt
google-generativeai>=0.3.0
openpyxl>=3.1.0
pandas>=2.0.0
reportlab>=4.0.0
matplotlib>=3.7.0
lxml>=4.9.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
